+++ pdata/plog/plogotlp/logs.go
@@ -126,9 +126,6 @@ type GRPCClient interface {
 	Export(ctx context.Context, request Request, opts ...grpc.CallOption) (Response, error)
 }
 
-// Deprecated: [0.61.0] Use GRPCClient instead
-type Client = GRPCClient
-
 type logsClient struct {
 	rawClient otlpcollectorlog.LogsServiceClient
 }
@@ -152,13 +149,13 @@ type GRPCServer interface {
 	Export(context.Context, Request) (Response, error)
 }
 
-// Deprecated: [0.61.0] Use GRPCServer instead
-type Server = GRPCServer
-
-// RegisterServer registers the Server to the grpc.Server.
-func RegisterServer(s *grpc.Server, srv GRPCServer) {
+// RegisterGRPCServer registers the Server to the grpc.Server.
+func RegisterGRPCServer(s *grpc.Server, srv GRPCServer) {
 	otlpcollectorlog.RegisterLogsServiceServer(s, &rawLogsServer{srv: srv})
 }
+
+// Deprecated: [0.62.0] Use RegisterGRPCServer instead
+var RegisterServer = RegisterGRPCServer
 
 type rawLogsServer struct {
 	srv GRPCServer